+++ lib/ecto_mysql_extras.ex
@@ -14,7 +14,13 @@ defmodule EctoMySQLExtras do
 
   @type repo() :: module() | {module(), node()}
 
-  @check_database [:db_settings, :db_status, :dirty_pages_ratio, :long_running_queries]
+  @check_database [
+    :db_settings,
+    :db_status,
+    :dirty_pages_ratio,
+    :long_running_queries,
+    :waits_for_checkpoint
+  ]
 
   @spec queries(repo()) :: map()
   def queries(_repo \\ nil) do
@@ -30,7 +36,8 @@ defmodule EctoMySQLExtras do
       table_size: EctoMySQLExtras.TableSize,
       total_index_size: EctoMySQLExtras.TotalIndexSize,
       total_table_size: EctoMySQLExtras.TotalTableSize,
-      unused_indexes: EctoMySQLExtras.UnusedIndexes
+      unused_indexes: EctoMySQLExtras.UnusedIndexes,
+      waits_for_checkpoint: EctoMySQLExtras.WaitsForCheckpoint
     }
   end
 
@@ -127,6 +134,9 @@ defmodule EctoMySQLExtras do
 
   @spec unused_indexes(repo(), keyword()) :: :ok | MyXQL.Result.t()
   def unused_indexes(repo, opts \\ []), do: query(:unused_indexes, repo, opts)
+
+  @spec waits_for_checkpoint(repo(), keyword()) :: :ok | MyXQL.Result.t()
+  def waits_for_checkpoint(repo, opts \\ []), do: query(:waits_for_checkpoint, repo, opts)
 
   defp default_opts(opts, nil), do: default_opts(opts, [])